Purpose: Bug fix
Description: On Cray T94 test_repack failed while copying
             chunked dataset. USHORT_MAX was assigned to maxchunk
             variable of int type. Since short is 8 bytes,
             maxchunk became negative -1 screwing the rest of the
             code.

Solution: Use INT_MAX instead of USHORT_MAX

Platforms tested: shanti with 32 and 64-bit, subzero (t94), lemieux (osf1)

Misc. update:

+++ mfhdf/hrepack/hrepack_sds.c
@@ -297,7 +297,7 @@ int copy_sds(int32 sd_in,
  if ( options->trip>0 ) 
  {
   int count=1, nchunks;
-  int maxchunk=USHRT_MAX;
+  int maxchunk=INT_MAX;
   if ( (chunk_flags == HDF_CHUNK) || (chunk_flags == (HDF_CHUNK | HDF_COMP)) )
   {
    for (j = 0; j < rank; j++) {
